+++ src/django/api/reports/facility_list_items_uploaded.sql
@@ -5,8 +5,8 @@ SELECT
   CASE WHEN u.email LIKE '%openapparel.org%' THEN 'y' ELSE 'n' END AS is_public_list,
   COUNT(*) AS item_count
   FROM api_facilitylistitem i
-         JOIN api_facilitylist l on i.facility_list_id = l.id
-         JOIN api_contributor c ON l.contributor_id = c.id
+         JOIN api_source s on i.source_id = s.id
+         JOIN api_contributor c ON s.contributor_id = c.id
          JOIN api_user u ON u.id = c.admin_id
  WHERE date_part('month', i.created_at) != date_part('month', now())
  GROUP BY date_part('month', i.created_at), is_public_list