+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 ### Changed
+- Implement Source model step 3 [#858](https://github.com/open-apparel-registry/open-apparel-registry/pull/858)
 
 ### Deprecated